Connectors: stop configuring service paths over the network.
The current scheme looks like a disaster waiting to happen. Service
paths are now specified at connector creation time and are in charge
of specifying it when they create request values. Webs_cli.service_path
is added to specify them on the cli.

It's now a bit easier to misconfigure gateways. One needs to keep say
nginx conf in sync with the service cli invocation, but it seems there
is no good answer to this problem (doing everything relative in the
service and its responses falls short once you want to define
a cookie path for your service).